Refactoring copy over
