+++ includes/home-endpoints.php
@@ -64,14 +64,6 @@ class DT_Webform_Home_Endpoints
         $namespace = 'dt-public/v' . $version;
 
         register_rest_route(
-            $namespace, '/webform/site_link_check', [
-                [
-                    'methods'  => WP_REST_Server::CREATABLE,
-                    'callback' => [ $this, 'site_link_check' ],
-                ],
-            ]
-        );
-        register_rest_route(
             $namespace, '/webform/transfer_collection', [
                 [
                 'methods'  => WP_REST_Server::READABLE,
@@ -79,28 +71,6 @@ class DT_Webform_Home_Endpoints
                 ],
             ]
         );
-    }
-
-    /**
-     * Verify is site is linked
-     *
-     * @param  WP_REST_Request $request
-     *
-     * @access public
-     * @since  0.1.0
-     * @return string|WP_Error|array The contact on success
-     */
-    public function site_link_check( WP_REST_Request $request )
-    {
-        $params = $request->get_params();
-
-        $prefix = 'dt_webform_site';
-
-        if ( isset( $params['id'] ) && isset( $params['token'] ) ) {
-            return DT_Webform_Api_Keys::check_api_key( $params['id'], $params['token'], $prefix );
-        } else {
-            return new WP_Error( "site_check_error", "Malformed request", [ 'status' => 400 ] );
-        }
     }
 
     /**
@@ -112,7 +82,7 @@ class DT_Webform_Home_Endpoints
     public function transfer_collection( WP_REST_Request $request ) {
 
         $params = $request->get_params();
-        $test = DT_Webform_Admin::verify_param_id_and_token( $params );
+        $test = DT_Api_Keys::verify_param_id_and_token( $params );
 
         if ( ! is_wp_error( $test ) && $test ) {
             if ( isset( $params['selected_records'] ) && ! empty( $params['selected_records'] ) ) {
