+++ includes/post-type-active-forms.php
@@ -452,7 +452,7 @@ class DT_Webform_Active_Form_Post_Type
         'name'        => __( 'Token', 'dt_webform' ),
         'description' => '',
         'type'        => 'display_only',
-        'default'     => DT_Webform_Api_Keys::generate_token( 16 ),
+        'default'     => DT_Api_Keys::generate_token( 16 ),
         'section'     => 'info',
         ];
