+++ includes/post-type-new-leads.php
@@ -110,8 +110,8 @@ class DT_Webform_New_Leads_Post_Type
     public function auto_accept( $post_id ) {
         $options = get_option( 'dt_webform_options' );
 
-        if ( ! DT_Webform_Admin::is_sites_keys_set() ) {
-            DT_Webform_Admin::set_auto_approve_to_false();
+        if ( ! DT_Webform_Settings::sites_keys_set() ) {
+            DT_Webform_Settings::set_auto_approve_to_false();
             $options['auto_approve'] = false;
         }
         elseif ( isset( $options['auto_approve'] ) && $options['auto_approve'] ) { // if auto approve is set
