+++ includes/remote-endpoints.php
@@ -63,14 +63,6 @@ class DT_Webform_Remote_Endpoints
         $namespace = 'dt-public/v' . $version;
 
         register_rest_route(
-            $namespace, '/webform/site_link_check', [
-                [
-                    'methods'  => WP_REST_Server::CREATABLE,
-                    'callback' => [ $this, 'site_link_check' ],
-                ],
-            ]
-        );
-        register_rest_route(
             $namespace, '/webform/form_submit', [
                 [
                     'methods'  => WP_REST_Server::CREATABLE,
@@ -78,37 +70,6 @@ class DT_Webform_Remote_Endpoints
                 ],
             ]
         );
-        register_rest_route(
-            $namespace, '/webform/custom_css', [
-                [
-                    'methods'  => WP_REST_Server::READABLE,
-                    'callback' => [ $this, 'custom_css' ],
-                ],
-            ]
-        );
-    }
-
-
-    /**
-     * Verify is site is linked
-     *
-     * @param  WP_REST_Request $request
-     *
-     * @access public
-     * @since  0.1.0
-     * @return string|WP_Error|array The contact on success
-     */
-    public function site_link_check( WP_REST_Request $request )
-    {
-        $params = $request->get_params();
-
-        $prefix = 'dt_webform_site';
-
-        if ( isset( $params['token'] ) ) {
-            return DT_Webform_Api_Keys::check_api_key( $params['id'], $params['token'], $prefix );
-        } else {
-            return new WP_Error( "site_check_error", "Malformed request", [ 'status' => 400 ] );
-        }
     }
 
     /**
@@ -155,24 +116,6 @@ class DT_Webform_Remote_Endpoints
         }
     }
 
-    /**
-     * Verify is site is linked
-     *
-     * @param  WP_REST_Request $request
-     *
-     * @access public
-     * @since  0.1.0
-     * @return string|WP_Error|array The contact on success
-     */
-    public function custom_css( WP_REST_Request $request )
-    {
-        $params = $request->get_params();
 
-        if ( isset( $params['token'] ) ) {
-            return DT_Webform_Remote::get_custom_css( $params['token'] );
-        } else {
-            return new WP_Error( "site_check_error", "Malformed request", [ 'status' => 400 ] );
-        }
-    }
 }
